+++ src/Range.php
@@ -11,38 +11,50 @@ namespace DecodeLabs\Compass;
 
 use Brick\Math\BigInteger;
 use DecodeLabs\Exceptional;
+use DecodeLabs\Fluidity\SingleParameterFactory;
+use DecodeLabs\Fluidity\SingleParameterFactoryTrait;
 use DecodeLabs\Glitch\Dumpable;
-use Stringable;
 
+/**
+ * @implements SingleParameterFactory<Ip|Scope|string|int|BigInteger>
+ */
 class Range implements
+    SingleParameterFactory,
     Scope,
-    Stringable,
     Dumpable
 {
+    /**
+     * @use SingleParameterFactoryTrait<Ip|Scope|string|int|BigInteger>
+     */
+    use SingleParameterFactoryTrait;
+
     use ScopeTrait;
 
     protected Ip $start;
     protected Ip $end;
+
+    protected ?Scope $originalScope = null;
 
     /**
      * Parse input value to Range
      */
     public static function parse(
         Ip|Scope|string|int|BigInteger $range
-    ): Range {
-        if ($range instanceof Range) {
+    ): static {
+        if ($range instanceof static) {
             return $range;
         }
 
-        return new self($range);
+        return new static($range);
     }
 
     /**
      * Init with input value
+     *
+     * @param Ip|Scope|string|int|BigInteger $range
      */
-    public function __construct(
-        Ip|Scope|string|int|BigInteger $range
-    ) {
+    public function __construct(mixed $range)
+    {
         // Single IP
         if (
             is_int($range) ||
@@ -60,23 +72,21 @@ class Range implements
         }
 
 
+        // CIDR
+        if (
+            is_string($range) &&
+            false !== strpos($range, '/')
+        ) {
+            $range = Block::parse($range);
+        }
+
+
         // Scope
         if ($range instanceof Scope) {
             $this->setRange(
                 $range->getFirstIp(),
-                $range->getLastIp()
-            );
-            return;
-        }
-
-
-        // CIDR
-        if (false !== strpos($range, '/')) {
-            $block = Block::parse($range);
-
-            $this->setRange(
-                $block->getFirstIp(),
-                $block->getLastIp()
+                $range->getLastIp(),
+                $range
             );
             return;
         }
@@ -112,10 +122,12 @@ class Range implements
 
     protected function setRange(
         Ip $start,
-        Ip $end
+        Ip $end,
+        ?Scope $originalScope = null
     ): void {
         $this->start = $start;
         $this->end = $end;
+        $this->originalScope = $originalScope;
 
         if ($this->start->isGreaterThan($this->end)) {
             throw Exceptional::UnexpectedValue(
@@ -264,6 +276,10 @@ class Range implements
      */
     public function __toString(): string
     {
+        if ($this->originalScope !== null) {
+            return $this->originalScope->__toString();
+        }
+
         return $this->start . '-' . $this->end;
     }