+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.1.2 (2022-09-15)
+* Use Fluidity SingleParameterFactory
+* Improved object instancing
+* Improved Range string conversion
+* Removed Dictum dependency
+
 ## v0.1.1 (2022-09-09)
 * Added isV* methods to Scopes